Switch from Jekyll to Sinatra
We want to be able to have events, posts, people and groups coming from
different sources and then being combined in the matrix. We also want to
persist the resulting memberships to a database so that EveryPolitican
can pull those memberships in from an API.

This first step just moves things into a Sinatra structure, but still
works in exactly the same way as the Jekyll version did.