Don't error if there is no group (party) set on a membership
At the moment the membership matrix assumes that parties are set on
memberships with on_behalf_of, but this isn't the case for some data
sources which have a separate membership to indicate which party someone
was in.  For the moment, it's helpful if we can still view these, even
if there will be no party information shown.