Add some basic tests using Mocha / Chai / PhantomJS
As membership_matrix becomes more capable it's feeling risky to add new
features without adding automated tests, but there was no test framework
in place.  This commit adds such a test framework, Mocha, set up to run
tests in the browser.  We're using Chai as the assertions library, and
headless testing (i.e. from the command-line) is done using PhantomJS.

This commit separates the immediately executed code that sets up the
membership matrix on the page (still in main.js) from the helper
functions that that code uses (in a new lib.js file).  This is so that
we don't get console errors when running the tests locally when it
fails to load the Popolo JSON or tries to draw the matrix on the page.